backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.36
asyncpg>=0.29.0
pydantic-settings>=2.2.1
httpx>=0.27.0
aiofiles>=23.2.1
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.1
python-multipart>=0.0.9
